ACTUALIZACIONES, SE QUITA EN MODEL CLIENTE EL DIA

+++ PIGGISWS/Interfaces/IProductoService.cs
@@ -7,5 +7,5 @@ public interface IProductoService
 
     Task<ServiceResponse<object>> GetProductosxAgente(int agente);
     Task<ServiceResponse<object>> GetTopProductosxAgente(int agente);
-    Task<ServiceResponse<object>> GetDescuentosxProductoAsync(int cproducto, decimal lprecios, decimal ccliente);
+    Task<ServiceResponse<object>> GetDescuentosxProductoAsync(decimal cproducto, decimal lprecios, decimal ccliente);
 }
